@@ -1,47 +1,41 @@
 {
-  "name": "little-brothers",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "npm run development",
-    "development": "mix",
-    "watch": "mix watch",
-    "watch-poll": "mix watch -- --watch-options-poll=1000",
-    "hot": "mix watch --hot",
-    "prod": "npm run production",
-    "production": "mix --production"
-  },
-  "dependencies": {
-    "@babel/polyfill": "^7.11.5",
-    "@fortawesome/fontawesome-free": "^5.15.4",
-    "@popperjs/core": "^2.10.2",
-    "bootstrap-vue": "^2.17.3",
-    "core-js": "^3.18.1",
-    "mutationobserver-shim": "^0.3.7",
-    "popper.js": "^1.16.1",
-    "portal-vue": "^2.1.7",
-    "v-calendar": "^2.3.4",
-    "vue": "^2.6.14",
-    "vue-msal": "^3.2.0",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^6.2.2",
-    "laravel-mix": "^6.0.41",
-    "prettier": "2.5.1",
-    "sass": "^1.43.3",
-    "sass-loader": "^10.0.2",
-    "vue-cli-plugin-bootstrap-vue": "~0.8.2",
-    "vue-template-compiler": "^2.6.11"
-  }
+    "name": "little-brothers",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "npm run development",
+        "development": "mix",
+        "watch": "mix watch",
+        "watch-poll": "mix watch -- --watch-options-poll=1000",
+        "hot": "mix watch --hot",
+        "prod": "npm run production",
+        "production": "mix --production",
+        "test:unit": "vue-cli-service test:unit",
+        "lint": "vue-cli-service lint"
+    },
+    "dependencies": {
+        "@babel/polyfill": "^7.11.5",
+        "@fortawesome/fontawesome-free": "^5.15.4",
+        "core-js": "^3.6.5",
+        "mutationobserver-shim": "^0.3.7",
+        "popper.js": "^1.16.1",
+        "register-service-worker": "^1.7.1",
+        "vue": "^3.2.29",
+        "vue-loader": "^17.0.0",
+        "vue-msal": "^3.2.0",
+        "vue-router": "^4.0.12",
+        "vue-template-compiler": "^2.6.12",
+        "vuex": "^3.2.29"
+    },
+    "devDependencies": {
+        "@popperjs/core": "^2.11.2",
+        "axios": "^0.21",
+        "bootstrap": "^5.1.3",
+        "laravel-mix": "^6.0.6",
+        "lodash": "^4.17.19",
+        "postcss": "^8.1.14",
+        "resolve-url-loader": "^3.1.2",
+        "sass": "^1.43.3",
+        "sass-loader": "^11.0.1"
+    }
 }
