@@ -1,5 +1,6 @@
 <script>
 import NavbarLink from './NavbarLink'
+
 export default {
   props: {},
   components: { NavbarLink },
@@ -12,7 +13,7 @@ export default {
 
 <template>
   <ul class="navbar" id="collapsable-navbar">
-      <img class="img" contain alt="LBFE logo" src="../../../../public/images/LBFE_LOGOV_red_upmichigan.png" />
+      <img class="img" contain alt="LBFE logo" src="../../images/LBFE_LOGOV_red_upmichigan.png" />
 
       <!-- Navbar - Admin Items -->
       <form v-if="$route.meta.admin">
@@ -35,7 +36,7 @@ export default {
   </ul>
 </template>
 
-<style scoped> 
+<style> 
   .navbar {
     color: var(--nb-text-color);
     background-color: var(--nb-color);
