@@ -1,106 +1,76 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-
-import DriverAbout from '../views/driver/DriverAbout.vue';
-import DriverHome from '../views/driver/DriverHome.vue';
-import DriverCalendar from '../views/driver/DriverCalendar.vue';
-import DriverUpdate from '../views/driver/DriverUpdate.vue';
-import DriverPrevUpdates from '../views/driver/DriverPrevUpdates.vue';
-import DriverMoreInfo from '../views/driver/DriverMoreInfo.vue';
-import Login from '../views/Login.vue';
-import DesktopCalendar from '../views/desktop/DesktopCalendar.vue';
-import AddEditAppointment from '../views/desktop/AddEditAppointment.vue';
-import ReportLog from '../views/desktop/ReportLog.vue';
-import UpdateLog from '../views/desktop/UpdateLog.vue';
-import Clients from '../views/desktop/Clients.vue';
-import DriverList from '../views/desktop/DriverList.vue'
-
-Vue.use(VueRouter);
+import { createRouter, createWebHistory } from 'vue-router'
 
 const routes = [
   {
     path: '/',
     name: 'Login',
-    component: Login,
+    component: () => import('../../views/Login.vue')
   },
   {
     path: '/appointment',
     name: 'AddEditAppointment',
-    component: AddEditAppointment,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/AddEditAppointment.vue')
   },
   {
     path: '/calendar',
     name: 'DesktopCalendar',
-    component: DesktopCalendar,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/DesktopCalendar.vue')
   },
   {
     path: '/reports',
     name: 'ReportLog',
-    component: ReportLog,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/ReportLog.vue')
   },
   {
     path: '/updates',
     name: 'UpdateLog',
-    component: UpdateLog,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/UpdateLog.vue')
   },
   {
     path: '/clients',
     name: 'Clients',
-    component: Clients,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/Clients.vue')
   },
   {
     path: '/drivers',
     name: 'DriverList',
-    component: DriverList,
-    meta: { admin: true },
+    component: () => import('../../views/desktop/DriverList.vue')
   },
   {
     path: '/driver',
     name: 'DriverHome',
-    component: DriverHome,
-    meta: { driver: true },
+    component: () => import('../../views/driver/DriverHome.vue')
   },
   {
     path: '/driver/about',
     name: 'DriverAbout',
-    component: DriverAbout,
-    meta: { driver: true },
+    component: () => import('../../views/driver/DriverAbout.vue')
   },
   {
     path: '/driver/calendar',
     name: 'DriverCalendar',
-    component: DriverCalendar,
-    meta: { driver: true },
+    component: () => import('../../views/desktop/DesktopCalendar.vue')
   },
   {
     path: '/driver/moreInfo',
     name: 'DriverMoreInfo',
-    component: DriverMoreInfo,
-    meta: { driver: true },
+    component: () => import('../../views/driver/DriverMoreInfo.vue')
   },
   {
     path: '/driver/update',
     name: 'DriverUpdate',
-    component: DriverUpdate,
-    meta: { driver: true },
+    component: () => import('../../views/driver/DriverUpdate.vue')
   },
   {
     path: '/driver/prev-updates',
     name: 'DriverPrevUpdates',
-    component: DriverPrevUpdates,
-    meta: { driver: true },
+    component: () => import('../../views/driver/DriverPrevUpdates.vue')
   },
 ];
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
-  routes,
-});
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
+  routes
+})
 
-export default router;
+export default router
