@@ -3,7 +3,6 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ItemController;
-use App\Http\Controllers\AppointmentController;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,14 +21,8 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::get('/items', [ItemController::class, 'index']);
 Route::prefix('/item')->group(function() {
-  Route::post('/store', [ItemController::class, 'store']);
+  //Route::post('/store', [ItemController::class, 'store']);
   Route::put('/{id}', [ItemController::class, 'update']);
   Route::delete('/{id}', [ItemController::class, 'destroy']);
 });
-
-Route::get('/appointments', [ItemController::class, 'index']);
-Route::prefix('/appointment')->group(function() {
-  Route::post('/store', [ItemController::class, 'store']);
-  Route::put('/{id}', [ItemController::class, 'update']);
-  Route::delete('/{id}', [ItemController::class, 'destroy']);
-});
+?>
